app/requirements.txt
psycopg[binary]>=3.1
requests>=2.31
tenacity>=8.2
python-dotenv>=1.0
